rename: change filename (libs/shared => libs/_shared) for distinction
