+++ app/(main)/page.tsx
@@ -1,12 +1,11 @@
 import { notFound } from 'next/navigation'
 
-import { getPagesHelper } from '@/libs/shared/apis/getNotion.api'
+import { getPagesHelper } from '@/libs/_shared/apis/getNotion.api'
 import {
   PagesHelperResponse,
   PartialPageObjectResponseMore,
-} from '@/libs/shared/types/responses.type'
-import { HomePageProps } from '@/libs/shared/types/routers.type'
-
+} from '@/libs/_shared/types/responses.type'
+import { HomePageProps } from '@/libs/_shared/types/routers.type'
 import HeroMain from '@/libs/heroMain/feature/heroMain.feature'
 import HeroMainArchives from '@/libs/heroMain/feature/heroMainArchives.feature'
 import HeroMainProjects from '@/libs/heroMain/feature/heroMainProjects.feature'
