+++ app/api/notion/content/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server'
 
-import { getPageContent } from '@/libs/shared/helpers/notion.helper'
+import { getPageContent } from '@/libs/_shared/helpers/notion.helper'
 
 export async function GET(req: NextRequest) {
   const detailId = req.nextUrl.searchParams.get('detailId') || ''
