+++ libs/_shared/_deprecated/components/post/data-access/extractHeadingsId.data-access.ts
