+++ libs/_shared/_deprecated/types/posts.type.ts
