+++ libs/_shared/styles/global.scss
