+++ libs/_shared/types/components.type.ts
